In order to reuse order details, we have to modify the CartRow component to display what we'd like
