+++ client/components/Cart/Cart.js
@@ -2,13 +2,13 @@ import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {fetchCart, getWines} from '../../store'
 import {Link} from 'react-router-dom'
+import OrderDetails from '../OrderDetails'
 
 export class Cart extends Component {
   componentDidMount() {
     this.props.getWines()
   }
   render() {
-    const cartWines = this.props.allWines
     const cart = Object.entries(this.props.cart).map(([key, val]) => {
       const wineInfo = this.props.allWines.filter(wine => +wine.id === +key)[0]
       return {...wineInfo, quantity: val}
@@ -17,26 +17,7 @@ export class Cart extends Component {
     const total = prices.reduce((tot, x) => tot + x, 0)
     return (
       <div id="display-cart">
-        <h2>Your Shopping Cart:</h2>
-        <table>
-          <thead>
-            <tr>
-              <th>Item</th>
-              <th>Price</th>
-              <th>Quantity</th>
-            </tr>
-          </thead>
-          <tbody>
-            {cart.map(item => (
-              <tr key={item.id}>
-                <td>{item.name}</td>
-                <td>{item.price}</td>
-                <td>{item.quantity}</td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-        <h3>Total: {total}</h3>
+        <OrderDetails wines={cart} total={total} />
         <Link to="/checkout">
           <button type="submit">Checkout</button>
         </Link>
